cloudrun/acta-engine/requirements.txt
flask==3.0.3
gunicorn==22.0.0
Pillow==10.4.0
PyMuPDF==1.24.9
pyzbar==0.1.9
